Fixed add contributor history issue
This way we dont need the full git history. It just takes the last
commit. Which might not work well if a PR comes in with multiple commits
and the last one is not the one that added a user. But we can live with
that as the task will be to only create this one small change in the PR
anyways.